% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{attach_extra}
\alias{attach_extra}
\title{Attach additional OrgDb information to results objects within a list}
\usage{
attach_extra(res_list, config)
}
\arguments{
\item{res_list}{List of DESeqResults objects}

\item{config}{Full config object, at least containing config$orgdb}
}
\value{
List of same results objects, but each one with additional columns
attached as specified in the config
Ensure everything has the same gene IDs; if so then we'll grab just the
gene IDs from the first one to use as our keys.
}
\description{
This can also fill in NAs in columns specified in the config with rownames
(e.g., fill in missing symbols with Ensembl IDs)
}
