% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{collapse_names}
\alias{collapse_names}
\title{Function to collapse res_list names with ontologies}
\usage{
collapse_names(res_list, config, sep = "*")
}
\arguments{
\item{res_list}{list of DESeq2 results objects}

\item{config}{Config object}

\item{collapse}{character string to separate the res_list names}
}
\value{
vector of strings with collapsed res_list names
}
\description{
Function to collapse res_list names with ontologies
}
