% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collapseReplicates2}
\alias{collapseReplicates2}
\title{DESeq2::collapseReplicates, but also fix the first column}
\usage{
collapseReplicates2(object, groupby)
}
\arguments{
\item{object}{Object to collapse. Typically a DESeq2 dds object}

\item{groupby}{Factor to group by. Typically a column from dds indicating
biological replicate (e.g., dds$biorep)}
}
\description{
DESeq2::collapseReplicates returns an object whose colData contains the
column used for collapsing, but only the first unique value of a collapsed
group is returned. This function makes the first column the same as the
rownames. This in turn allows the colData to meet expectations of other
lcdbwf functions and play nicer with dplyr.
}
