% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collect_objects}
\alias{collect_objects}
\title{Find objects whose names match a pattern and store in a list}
\usage{
collect_objects(pattern, fixed = FALSE)
}
\arguments{
\item{pattern}{Pattern to provide to grepl}

\item{fixed}{Passed to grepl()}
}
\value{
List of identified objects
}
\description{
The pattern will be used to search within the global environment. The names
of the list will be the variable names with the matched pattern removed.
}
