% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compose_results}
\alias{compose_results}
\title{Compose an RDS file to be used in downstream tools.}
\usage{
compose_results(res_list, dds_list, enrich_list)
}
\arguments{
\item{res_list}{List of results objects and associated metadata. See details
for format.}

\item{dds_list}{List of dds objects used in results}

\item{enrich_list}{List of enrichment results objects. See details for format.}
}
\description{
Compose an RDS file to be used in downstream tools.
}
\details{
res_list and dds_list are required. \code{res_list} has the following format.

list(
ko.vs.wt=list(             # names of the list are short keys
res=DESeqResults object, # standard results object, may have columns for SYMBOL, etc
dds="main",              # string key into dds_list
label="KO vs WT"),       # Description of contrast
het.vs.wt=list(
....
),
...
)

Note that values of "dds" in the res_list's items must be names in
\code{dds_list}. In this example, there is only one, "main". \code{dds_list} has the
following format:

list(
main=DESeqDataSet,
...
)

\code{enrich_list} is optional. Note that \code{enrich_list}, if provided, must use
results names available in \code{res_list}. In this example, the names are
"ko.vs.wt" and "het.vs.wt". It has the following format:

list(
ko.vs.wt=list(
up=list(
BP=enrichResults object,
CC=enrichResults object,
MF=enrichResults object,
...
),
down=list(
BP=enrichResults object,
...
),
het.vs.wt=list(
up=list(...),
down=list(...)
),
...
)
}
