% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{counts.plot}
\alias{counts.plot}
\title{Plot genes' normalized counts across samples}
\usage{
counts.plot(df, rank.nb = NULL, no.aes = FALSE, facet = "label")
}
\arguments{
\item{df}{dataframe from which to extract genes normalized_counts}

\item{rank.nb}{rank number less or equal used to filter the genes to be plotted}

\item{no.aes}{whether to include the default aes or return the ggplot object
without aes}

\item{facet}{name of column to use for faceting}
}
\value{
ggplot object
}
\description{
Plot genes' normalized counts across samples
}
