% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{dds_coefs}
\alias{dds_coefs}
\title{Prepare coefficients that can be combined for more complex models.}
\usage{
dds_coefs(dds, ..., expand = FALSE)
}
\arguments{
\item{dds}{DESeqDataSet object or string. If string, looks in the global
environment for "dds_list" and takes the dds object with this name in the
list.}

\item{...}{Additional filtering criteria which are handed off to
dplyr::filter and used to filter the colData(dds).}
}
\value{
A named vector, representing the coefficients of the samples
selected via \code{...}
}
\description{
Heavily influenced by
https://github.com/tavareshugo/tutorial_DESeq2_contrasts/blob/main/DESeq2_contrasts.md
}
\details{
First we simulate some data to work with:

dds <- makeExampleDESeqDataSet(n = 1000, m = 9, betaSD = 2)
dds$condition <- NULL
dds$colour <- factor(rep(c("pink", "yellow", "white"), each = 3))
dds$colour <- relevel(dds$colour, "white")
design(dds) <- ~colour

Here's what the colData looks like:
colData(dds)
\subsection{DataFrame with 9 rows and 1 column}{
}

\subsection{colour}{
}

\subsection{factor>}{
}

\subsection{sample1   pink}{
}

\subsection{sample2   pink}{
}

\subsection{sample3   pink}{
}

\subsection{sample4   yellow}{
}

\subsection{sample5   yellow}{
}

\subsection{sample6   yellow}{
}

\subsection{sample7   white}{
}

\subsection{sample8   white}{
}

\subsection{sample9   white}{

If we wanted a contrast for "pigmented vs white", we could then do:

pigmented <- dds_coefs(dds, colour \%in\% c('pink', 'yellow'))
white <- dds_coefs(dds, colour=='white')
res <- results(dds, contrast=pigmented-white)
}
}
