% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R
\name{dds_diagnostics}
\alias{dds_diagnostics}
\title{Plot various diagnostics for dds objects.}
\usage{
dds_diagnostics(dds_list, text)
}
\arguments{
\item{dds_list}{List of dds objects}

\item{text}{Text config object}
}
\description{
Plot various diagnostics for dds objects.
}
