% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{DESeqDataSetFromCombinedFeatureCounts}
\alias{DESeqDataSetFromCombinedFeatureCounts}
\title{Load a single combined featureCounts table into a DESeq object.}
\usage{
DESeqDataSetFromCombinedFeatureCounts(
  filename,
  sampletable,
  design,
  sample_func = lcdbwf.samplename,
  subset_counts = FALSE,
  ...
)
}
\arguments{
\item{filename}{Filename containing featureCounts combined output}

\item{sampletable}{data.frame containing at least sample names as first
column}

\item{design}{Model used for creating DESeq object}

\item{sample_func}{Function that will be applied to each column name
to align it to the input sampletable. Takes a character vector as
input and returns a character vector of adjusted strings.}

\item{subset_counts}{If TRUE, then counts are subsetted by the provided
sampletable. That is, only the counts columns that match a rowname of
the provided sampletable are included. If FALSE (default), an error
is raised reporting the differences in sampletable and counts data.
It is always an error if counts data does not have an entry for
a sample in the sampletable.}
}
\value{
DESeq object

Additional args are passed to DESeq2::DESeqDataSetFromMatrix.
}
\description{
Load a single combined featureCounts table into a DESeq object.
}
