% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{DESeqDataSetFromFeatureCounts}
\alias{DESeqDataSetFromFeatureCounts}
\title{Load featureCounts output into a DESeq object.}
\usage{
DESeqDataSetFromFeatureCounts(
  sampleTable,
  directory = ".",
  design,
  ignoreRank = FALSE,
  ...
)
}
\arguments{
\item{sampleTable}{data.frame containing at least "featurecounts.path" column}

\item{directory}{Paths to featureCounts output are relative to this dir}

\item{design}{Model used for creating DESeq object}
}
\value{
DESeq object

Additional args are passed to DESeq2::DESeqDataSetFromMatrix.
}
\description{
Revised version of DESeq2::DESeqDataSetFromHTSeqCount to handle the
featureCounts default output format, which contains many more columns.
}
