% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{emapplots}
\alias{emapplots}
\title{Wrapper for enrichplot::emapplot}
\usage{
emapplots(
  enrich_res,
  config,
  name = NULL,
  direction = NULL,
  ont = NULL,
  truncate_to = 50
)
}
\arguments{
\item{enrich_res}{enrichResult object}

\item{config}{Config object}

\item{name, }{direction, ont These may be passed in when iterating over
nested list of results}

\item{truncate_to}{Truncate labels to this long}
}
\description{
Wrapper for enrichplot::emapplot
}
