% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exported_excel}
\alias{exported_excel}
\title{Export Excel file of results}
\usage{
exported_excel(res_list, dds_list, file = "results.xlsx")
}
\arguments{
\item{res_list}{Results list object}

\item{dds_list}{List of dds objects}

\item{file}{Output .xlxs file}
}
\description{
One contrast per worksheet, with normalized counts and filters enabled
}
