% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exported_tsvs}
\alias{exported_tsvs}
\title{Save TSVs to ourput directory, named after the names in the list.}
\usage{
exported_tsvs(res_list, directory = "results")
}
\arguments{
\item{res_list}{"reslist" object}

\item{directory}{Directory in which to save results}
}
\description{
Save TSVs to ourput directory, named after the names in the list.
}
