% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_chunk_names}
\alias{find_chunk_names}
\title{Find named chunks that match a pattern}
\usage{
find_chunk_names(pattern)
}
\arguments{
\item{pattern}{Regular expression used to find chunks.}
}
\value{
List of chunk names matching pattern
}
\description{
Useful for depending on chunks that are named a certain way.
}
