% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{folded_markdown}
\alias{folded_markdown}
\title{Print markdown enclosed by "details" HTML tags.}
\usage{
folded_markdown(text, summary)
}
\arguments{
\item{text}{Markdown text to print}

\item{summary}{Text to use next to dropdown arrow}
}
\description{
Print markdown enclosed by "details" HTML tags.
}
