% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format.group.name}
\alias{format.group.name}
\title{Compute label for one component of an arbitrary design matrix}
\usage{
\method{format}{group.name}(pattern.mat)
}
\arguments{
\item{pattern.mat}{single row from unique(model.matrix) call from a dds object}
}
\value{
formatted column name
}
\description{
Compute label for one component of an arbitrary design matrix
}
