% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplots.R
\name{fromList.with.names}
\alias{fromList.with.names}
\title{Prepare list for UpSet plots, but include rownames}
\usage{
fromList.with.names(lst)
}
\arguments{
\item{x}{List of sets to compare (same input as to UpSetR::fromList)}
}
\value{
data.frame of 1 and 0 showing which genes are in which sets
}
\description{
Prepare list for UpSet plots, but include rownames
}
