% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{genes_to_label}
\alias{genes_to_label}
\title{Identify genes to label in MA plots and volcano plots}
\usage{
genes_to_label(res, n = 5, config)
}
\description{
Default is to rank by log2FoldChange and return the top 5 and bottom 5.
}
