% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get.frac}
\alias{get.frac}
\title{Convert "1/100" to 0.01.}
\usage{
get.frac(x)
}
\arguments{
\item{x}{Character vector to convert}
}
\description{
clusterProfiler report columns that are strings of numbers; this converts to
a fraction
}
