% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{get.orgdb}
\alias{get.orgdb}
\title{Get the OrgDb for the specified organism, using the cached AnnotationHub.}
\usage{
get.orgdb(config)
}
\arguments{
\item{config}{List containing the following named items:
\itemize{
\item annotation_genus_species: e.g., "Homo sapiens"
\item hub_cache: Optional path, relative to the working directory, in which
to store the downloaded OrgDb database
\item annotation_key_override: Optional key that if specified overrides
annotation_genus_species. Use this to be exact about the version of the
OrgDb you want to use.
}}
}
\value{
OrgDb object
}
\description{
Get the OrgDb for the specified organism, using the cached AnnotationHub.
}
