% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.sig}
\alias{get.sig}
\title{Return index of up/down/changed genes}
\usage{
get.sig(
  x,
  direction = "up",
  alpha = 0.1,
  lfc.thresh = 0,
  return.names = TRUE,
  column = NULL
)
}
\arguments{
\item{x}{DESeq2 results object, or data.frame created from one}

\item{direction}{Direction in 'up', 'dn', 'down', 'ch', 'changed'}

\item{alpha}{FDR lower than this will be considered significant}

\item{return.names}{If TRUE, returns the rownames of selected genes; if
FALSE return boolean index of length(x)}

\item{column}{If provided and return.names is TRUE, then return this column
instead of the default rownames}

\item{thresh}{Log2 fold change threshold. If e.g. 2, will return < -2 and/or2, depending on the value of "direction"}
}
\value{
Character vector of rownames (if return.names=TRUE) or boolean vector of genes selected.
}
\description{
Return index of up/down/changed genes
}
