% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{get_annotation_db}
\alias{get_annotation_db}
\title{Get the OrgDb for the specified organism, using the cached AnnotationHub.}
\usage{
get_annotation_db(
  config,
  dbtype,
  genus_species = NULL,
  orgdb_key_override = NULL,
  txdb_key_override = NULL,
  cache = NULL
)
}
\arguments{
\item{config}{List containing the following named items:
\itemize{
\item annotation_genus_species: e.g., "Homo sapiens"
\item hub_cache: Optional path, relative to the working directory, in which
to store the downloaded OrgDb database
\item annotation_key_override: Optional key that if specified overrides
annotation_genus_species. Use this to be exact about the version of the
OrgDb you want to use.
}}
}
\value{
OrgDb object
}
\description{
Get the OrgDb for the specified organism, using the cached AnnotationHub.
}
