% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{get_annotation_hub}
\alias{get_annotation_hub}
\title{Get the AnnotationHub object, using settings from config}
\usage{
get_annotation_hub(config, localHub = NULL, force = NULL, cache = NULL)
}
\arguments{
\item{config}{Config object}

\item{localHub, }{force, cache Arguments passed on to AnnotationHub(); these
can be used to override what's in the config.}
}
\description{
Get the AnnotationHub object, using settings from config
}
