% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_config}
\alias{get_config}
\title{Get config object from global env}
\usage{
get_config()
}
\value{
Config object (nested list) if in global env, otherwise NULL.
}
\description{
Get config object from global env
}
