% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{get_dds}
\alias{get_dds}
\title{Helper function to get dds object from global env.}
\usage{
get_dds(dds)
}
\arguments{
\item{dds}{If string, then look for that name in \code{dds_list} in the global
env. Otherwise, return it as-is.}
}
\description{
Helper function to get dds object from global env.
}
