% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_go_descriptions}
\alias{get_go_descriptions}
\title{Extract all GO IDs and their respective descriptions.}
\usage{
get_go_descriptions()
}
\description{
Extract all GO IDs and their respective descriptions.
}
