% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_go_term2gene}
\alias{get_go_term2gene}
\title{Return a list of GO annotation dataframes, one per ontology.}
\usage{
get_go_term2gene(config)
}
\arguments{
\item{orgdb}{OrgDb}

\item{keytype}{Key type intended to be use with clusterProfiler::enricher
function.}
}
\value{
List of dataframes; list names are "BP", "CC", "MF".
}
\description{
Return a list of GO annotation dataframes, one per ontology.
}
