% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_go_term2gene_alt}
\alias{get_go_term2gene_alt}
\title{Alternative implementation of get_go_term2gene, stored here for historical
comparison purposes.}
\usage{
get_go_term2gene_alt(orgdb, keytype)
}
\description{
Alternative implementation of get_go_term2gene, stored here for historical
comparison purposes.
}
