% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_kegg_list}
\alias{get_kegg_list}
\title{Get KEGG pathway annotation information}
\usage{
get_kegg_list(config)
}
\arguments{
\item{config}{Config object}
}
\value{
list with term2gene and term2name data frames
}
\description{
Get KEGG pathway annotation information
}
