% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_kegg_species}
\alias{get_kegg_species}
\title{Get the species name that works with the KEGG database}
\usage{
get_kegg_species(config)
}
\arguments{
\item{config}{Config object}
}
\value{
KEGG-compatible species name
}
\description{
The KEGG pathway needs the species name in a
specific format, e.g. for 'Homo sapiens' the
KEGG version would be 'hsa'.
}
