% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_msigdb_df}
\alias{get_msigdb_df}
\title{Get the MSigDB data for the organism provided in the config.}
\usage{
get_msigdb_df(config)
}
\description{
Get the MSigDB data for the organism provided in the config.
}
