% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_msigdb_term2gene_list}
\alias{get_msigdb_term2gene_list}
\title{Return a list of MSigDB gene sets, one per subcategory}
\usage{
get_msigdb_term2gene_list(msigdb_df)
}
\description{
Names of the list are concatenated category_subcategory.
}
