% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_msigdb_term2name}
\alias{get_msigdb_term2name}
\title{TERM2NAME for MSigDB}
\usage{
get_msigdb_term2name(msigdb_df)
}
\arguments{
\item{msigdb_df}{data.frame of MSigDB for the species, likely from
get_msigdb_df() or msigdbr::msigdbr().}
}
\description{
TERM2NAME for MSigDB
}
