% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_ontology_list}
\alias{get_ontology_list}
\title{All-in-one function to get ontology information}
\usage{
get_ontology_list(config)
}
\arguments{
\item{config}{Config object.}
}
\value{
List of lists where top-level elements are ontology
names. Each element is a list with TERM2GENE & TERM2NAME
data frames
}
\description{
All-in-one function to get ontology information
}
