% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_sig}
\alias{get_sig}
\title{Return up/down/changed genes}
\usage{
get_sig(
  x,
  direction = "up",
  alpha = 0.1,
  lfc_thresh = 0,
  return_type = "rownames"
)
}
\arguments{
\item{x}{DESeq2 results object, or data.frame created from one}

\item{direction}{Direction in 'up', 'dn', 'down', 'ch', 'changed'}

\item{alpha}{FDR lower than this will be considered significant}

\item{return_type}{If \code{return_type} is "rownames", return the rownames of
the dataframe of the significant genes. If "results", return a subset of
the original object. If "bool", then return an index of TRUE/FALSE.}

\item{thresh}{Log2 fold change threshold. If e.g. 2, will return < -2
and/or > 2, depending on the value of "direction"}
}
\value{
Character vector of rownames, bool vector, or subset of original
results object depending on the value of \code{return_type}.
}
\description{
Return up/down/changed genes
}
