% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lcdbwf.samplename}
\alias{lcdbwf.samplename}
\title{Convert a vector of BAM filenames into a vector of samplenames.}
\usage{
lcdbwf.samplename(x)
}
\arguments{
\item{x}{Character vector of BAM filenames (e.g., from the header of
featureCounts)}
}
\description{
Convert a vector of BAM filenames into a vector of samplenames.
}
