% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lfc.filter}
\alias{lfc.filter}
\title{Filter results by log2FoldChange sign}
\usage{
lfc.filter(res, reverse = FALSE)
}
\arguments{
\item{res}{DESeq2 results object}

\item{reverse}{If TRUE then filter negative lfc}
}
\value{
filtered genes
}
\description{
Filter results by log2FoldChange sign
}
