% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lfc.order}
\alias{lfc.order}
\title{Re-order results by log2FoldChange}
\usage{
lfc.order(res, reverse = FALSE)
}
\arguments{
\item{res}{DESeq2 results object}

\item{reverse}{If TRUE then sort high-to-low}
}
\value{
Re-ordered results object
}
\description{
Re-order results by log2FoldChange
}
