% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_config}
\alias{load_config}
\title{Load config file}
\usage{
load_config(filename)
}
\arguments{
\item{filename}{YAML filename to load}
}
\value{
Nested list object containing configuration
}
\description{
Load config file
}
\details{
For testing purposes, if the environment variable LCDB_WF_TEST exists then
assume we are running a test and set config$toggle$test to TRUE. This will
override any settings in the config YAML. Otherwise, use what is in the
config, and if not specified in the config then default to FALSE.
}
