% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make.dds.list}
\alias{make.dds.list}
\title{Make a list of dds objects}
\usage{
make.dds.list(
  deseq_obj_list,
  salmon.files = NULL,
  combine.by = FALSE,
  remove.version = TRUE,
  ...
)
}
\arguments{
\item{deseq_obj_list}{A named list of lists. Each list is used as the first
argument to \code{make.dds}; see the documentation of that function for
details.}
}
\value{
A list of dds objects.
}
\description{
Helper function to construct a list of dds objects. The \code{make.dds} function
does all the work; this just sets up some sane defaults and does the map()
call.
}
