% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make.dds}
\alias{make.dds}
\title{Make a single dds object}
\usage{
make.dds(
  design_data,
  salmon.files = NULL,
  combine.by = NULL,
  remove.version = FALSE,
  ...
)
}
\arguments{
\item{design_data}{List of 2 to 4 items. The first item must be the
sampletable. The second item must be the design. The remaining
items are optional, named arguments: \code{file} is the featureCounts
file for all samples, and/or \code{args} is a list of arguments to be
passed to the constructor.}

\item{salmon.files}{If you want the dds objects to be generated from salmon
counts, then provide the txi object from tximport. Otherwise,
leave as NULL to use featureCounts. The value of this argument is
used for all dds objects in the returned list (that is, the
returned list cannot have a mix of salmon and featureCounts; if
you want both you'll need to call this function twice).}

\item{combine.by}{The column to collapse technical replicates by. Rows in
the sampletable that share the same value of this column will be
combined using DESeq2::collapseReplicates.}

\item{remove.version}{If TRUE, gene (or transcript) version information --
the ".1" in "ENSG0000102345.1" -- will be stripped off.}

\item{...}{Additional arguments will be passed on to the DESeq() call (e.g.,
parallel, fitType, etc)}
}
\description{
This function is intended to be applied to a list of design data.
}
