% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R
\name{make_dds}
\alias{make_dds}
\title{Make a single dds object}
\usage{
make_dds(
  design_data,
  config = NULL,
  salmon = NULL,
  collapse_by = NULL,
  strip_dotted_version = NULL,
  featureCounts = "../data/rnaseq_aggregation/featurecounts.txt",
  ...
)
}
\arguments{
\item{design_data}{Named list of 2 to 4 items. At least "sampletable" (the
colData, as a data.frame or tibble) and "design" (e.g., \code{~group}) are
required. The optional named items are "file", which is the featureCounts
file containing counts for all samples, and "args" which is a list of
arguments to be passed to the constructor (e.g.,
\verb{args=list(subset.counts=TRUE))}.}

\item{salmon}{Set to TRUE if you want to create dds objects from Salmon
counts (will be loaded via tximport) Otherwise, use featureCounts}

\item{collapse_by}{The column to collapse technical replicates by. Rows in
the sampletable that share the same value of this column will be combined
using DESeq2::collapseReplicates.}

\item{strip_dotted_version}{If TRUE, then remove Ensembl-style dotted
version numbers from gene IDs (ENSG000001.1 -> ENSG000001)}

\item{featureCounts}{Location of featureCounts output to be loaded}

\item{...}{Additional arguments will be passed on to the DESeq() call (e.g.,
parallel, fitType, etc)}
}
\description{
This function is intended to be applied to a list of design data along with
a config object to create a list of dds objects:
}
\details{
dds_list <- map(dds_list, make_dds, config)

While the intended use is to provide a config object, individual arguments
can be used to override individual config items.
}
