% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{make_results}
\alias{make_results}
\title{Convenience function for building contrasts using the data structures
expected by lcdbwf.
This function is useful for ensuring that the dds object used when creating
the results is the same one that is labeled, so that when the dds_list and
res_list are composed together things match up nicely.}
\usage{
make_results(dds_name, label, ...)
}
\arguments{
\item{dds_name}{String key into dds_list. Will be used to extract the actual
dds object to provide to results()}

\item{label}{Label to describe this contrast which will be used in headings.}

\item{...}{Additional arguments are passed to results() and lfcShrink().}

\item{contrast}{Any contrast supported by DESeq2::results and DESeq2::lfcShrink}
}
\value{
List with names "res" (the results object), "dds" (the string value
passed in that was used to retrieve the actual dds object used for the
results), and "label" (exactly as passed in).
}
\description{
This function first runs results and then runs lfcShrink on that results
object. Arguments in ... are passed through to results() and then to
lfcShrink(), using formals() to detect which function accepts which
arguments. Those arguments accepted by both (contrast, lfcThreshold,
format, saveCols, parallel, BPPARAM) are passed to both.
}
\details{
NOTE: this function expects \code{dds_list} and \code{config} objects to be available
in the global environment.

This function also adds the shrinkage type as an additional item to the
metadata of the results object.
}
