% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{match_from_dots}
\alias{match_from_dots}
\title{Return only the list items in \code{dots} that are arguments for \code{func}}
\usage{
match_from_dots(dots, func)
}
\arguments{
\item{dots}{List of arguments, most often created via \code{list(...)}}

\item{func}{Function that will be inspected for valid arguments}
}
\value{
List of arguments in \code{dots} that are valid arguments for \code{func}.
}
\description{
Thanks to
https://community.rstudio.com/t/dots-vs-arg-lists-for-function-forwarding/4995/2
for the idea.
}
