% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mdcat}
\alias{mdcat}
\title{Simple wrapper of cat() that makes markdown text easier to print}
\usage{
mdcat(...)
}
\arguments{
\item{...}{Arguments to print

Make sure you're in an R code chunk with \code{results="asis"} set.}
}
\description{
Simple wrapper of cat() that makes markdown text easier to print
}
