% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{modify_test_config}
\alias{modify_test_config}
\title{Modify config file for testing}
\usage{
modify_test_config(config)
}
\arguments{
\item{config}{Config object}
}
\description{
When running tests in resource-limited environments, we may need to adjust
the config file accordingly. Make all such adjustments in this function.
}
