% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{my.counts}
\alias{my.counts}
\title{Plot a gene's normalized counts across samples}
\usage{
my.counts(gene, dds, label = NULL, intgroup = "group")
}
\arguments{
\item{gene}{Gene ID}

\item{dds}{DESeq object from which to extract counts}
}
\value{
ggplot object
}
\description{
Plot a gene's normalized counts across samples
}
