% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{my.summary}
\alias{my.summary}
\title{Summarize DESeq2 results into a dataframe}
\usage{
my.summary(res, dds, alpha, lfc.thresh = 0, ...)
}
\arguments{
\item{res}{DESeq2 results object}

\item{dds}{DEseq2 object}

\item{alpha}{Alpha level at which to call significantly changing genes}
}
\value{
Dataframe of summarized results
}
\description{
summary(res) prints out info; this function captures it into a dataframe
}
