% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nested.lapply}
\alias{nested.lapply}
\title{lapply over a nested list}
\usage{
nested.lapply(x, subfunc, ...)
}
\arguments{
\item{x}{Nested list. Expected 2 levels deep, and each item is the same type}

\item{subfunc}{Function to apply over nested "leaf" items}

\item{...}{Additional arguments to be passed to subfunc}
}
\value{
Nested list with same structure but tranformed values
}
\description{
lapply over a nested list
}
