% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{padj.order}
\alias{padj.order}
\title{Re-order results by adjusted pval}
\usage{
padj.order(res, reverse = FALSE)
}
\arguments{
\item{res}{DESeq2 results object}

\item{reverse}{If TRUE then sort high-to-low}
}
\value{
Re-ordered results object
}
\description{
Re-order results by adjusted pval
}
