% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Plot a clustered heatmap of samples}
\usage{
\method{plot}{heatmap}(rld, colData, cols.for.grouping)
}
\arguments{
\item{rld}{DESeqTransform object, typically output from running rlog()}

\item{colData}{Dataframe of metadata, used for annotating heatmap}

\item{cols.for.grouping}{Columns in colData to annotate heatmap with}
}
\value{
matrix of sample distances
}
\description{
Plot a clustered heatmap of samples
}
