% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplots.R
\name{plot.upsets}
\alias{plot.upsets}
\title{Plot UpSet plots}
\usage{
\method{plot}{upsets}(res_list, label_column = NULL, alpha = 0.1, lfc.thresh = 0)
}
\arguments{
\item{res_list}{"reslist" object}

\item{label_column}{When creating output files, use this column to label genes}
}
\description{
Plot UpSet plots
}
