% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplots.R
\name{plot_upsets}
\alias{plot_upsets}
\title{Plot UpSet plots}
\usage{
plot_upsets(res_list, label_column = NULL, alpha = 0.1, lfc_thresh = 0)
}
\arguments{
\item{res_list}{"reslist" object}

\item{label_column}{When creating output files, use this column to label genes}
}
\description{
Plot UpSet plots
}
