% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_volcano_label}
\alias{plot_volcano_label}
\title{Plot a volcano plot labeled with selected genes}
\usage{
plot_volcano_label(
  res,
  fdr_thres = 0.1,
  fd_thres = 0,
  fc_lim = NULL,
  genes_to_label = NULL,
  label_column = NULL
)
}
\arguments{
\item{fdr_thres}{fdr_threshold for defining statistical significance}

\item{fd_thres}{log2FoldChange cutoff for defining statistical significance}

\item{fc_lim}{User-defined limits for x-axis (log2FC). If NULL, this is defined as
the (floor, ceiling) of range(res$log2FoldChange)}

\item{genes_to_label}{genes_to_label on the volcano plot. NULL, by default}

\item{res.list}{data.frame (or list) with log2FoldChange and padj columns (or elements).
Row names of the data.frame should be gene IDs/symbols and match the format of lab.genes}

\item{col}{Column of \code{res} in which to look for \code{genes_to_label}. If NULL,
rownames are used.}
}
\value{
Handle to ggplot
}
\description{
Plot a volcano plot labeled with selected genes
}
