% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pval.hist}
\alias{pval.hist}
\title{Plot a histogram of raw pvals}
\usage{
pval.hist(res)
}
\arguments{
\item{res}{DESeq2 results object}
}
\value{
Side effect is to create plot
}
\description{
Plot a histogram of raw pvals
}
