% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pval_hist}
\alias{pval_hist}
\title{Plot a histogram of raw pvals}
\usage{
pval_hist(res)
}
\arguments{
\item{res}{DESeq2 results object}
}
\value{
Side effect is to create plot
}
\description{
This is right out of the DESeq2 vignette, from the section about independent
filtering. The resulting histogram indicates pvals for those genes kept and
removed before multiple testing adjustment.
}
