% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{results_diagnostics}
\alias{results_diagnostics}
\title{Perform diagnostics on results objects and plot}
\usage{
results_diagnostics(res, dds, name, config, text)
}
\arguments{
\item{res}{Results object}

\item{dds}{Corresponding dds object used to generate results}

\item{config}{Config object}

\item{text}{Text config object, used for labeling plots}
}
\description{
This function runs some diagnostic functions on results objects which can be
helpful during troubleshooting odd behavior.
}
