% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rownames.first.col}
\alias{rownames.first.col}
\title{Move rownames of data.frame to first column}
\usage{
rownames.first.col(x, name = "names")
}
\arguments{
\item{x}{data.frame}

\item{name}{Name to use as the new first column}
}
\value{
data.frame with original rownames moved to new first column with
optional name. rownames on the new data.frame are set to NULL.
}
\description{
Move rownames of data.frame to first column
}
