% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{run_enricher}
\alias{run_enricher}
\title{Function to run enrichment on a res_list}
\usage{
run_enricher(res_list, ontology_list, config, cores = 1, sep = "*")
}
\arguments{
\item{res_list}{list of DESeq2 results objects}

\item{config}{Config object}

\item{cores}{Number of cores to run it on}

\item{sep}{Character to separate res_list names}
}
\value{
nested list of enrichResult objects
}
\description{
Function to run enrichment on a res_list
}
