% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{run_enrichment}
\alias{run_enrichment}
\title{All-in-one enrichment function.}
\usage{
run_enrichment(res, TERM2GENE, TERM2NAME, config, direction, kind = "OR", ...)
}
\arguments{
\item{res}{DESeq2 results object}

\item{TERM2GENE}{A data.frame, first column GO ID, second column gene name.
It is assumed that the gene names are the same format as those in
rownames(res).}

\item{TERM2NAME}{A data.frame, first column GO ID, second column description.}

\item{config}{Config object. pvalueCutoff and qvalueCutoff will be taken from here.}

\item{direction}{One of "up", "down", or "changed". Will use alpha and
lfc_thresh from the config.}

\item{kind}{One of "OR" for overrepresentation or "GSEA" for gene set
enrichment analysis.}

\item{...}{Additional arguments are passed on to enricher() for kind="OR" or
GSEA() for kind="GSEA".}
}
\value{
An enrichResults object from
}
\description{
Designed to not require an orgdb, and instead requires dataframes of
term2gene and optionally term2name.
}
