% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{run_functional_enrichment}
\alias{run_functional_enrichment}
\title{Run overrepresentation analysis}
\usage{
run_functional_enrichment(
  genes.df,
  orgdb,
  orgdb_config,
  truncate_labels_to = 50
)
}
\arguments{
\item{genes.df}{Nested list of lists, see details below.}

\item{orgdb}{OrgDb containing GO and KEGG terms}

\item{orgdb_config}{Config list (documented elsewhere)}

\item{truncate_labels_to}{Long labels will be truncated to this many
characters to make plots more readable}
}
\description{
Run overrepresentation analysis
}
\details{
\code{genes.df} is a nested list of lists. The first level represents contrasts
(condition1 vs condition2, etc) and within each contrast are the selections
(typically "up" and "dn"). The values are character vectors of gene IDs. For
example:

genes.df <- list(
contrast1=list(up=c('gene1', 'gene2'),
dn=c('gene8', 'gene9', 'gene10')
),
contrast2=list(...
)
)

\code{orgdb_config} is a nested list containing configuration information and is
documented elsewhere.
}
