% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sizefactors_barplot}
\alias{sizefactors_barplot}
\title{Barplot of size factors by sample}
\usage{
sizefactors_barplot(dds)
}
\arguments{
\item{dds}{DESeqDataSet object}
}
\description{
Barplot of size factors by sample
}
