% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sizefactors_vs_total}
\alias{sizefactors_vs_total}
\title{Scatterplot of size factors vs total read count}
\usage{
sizefactors_vs_total(dds)
}
\arguments{
\item{dds}{DESeqDataSet object}
}
\description{
Scatterplot of size factors vs total read count
}
