% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dds.R
\name{strip_dotted_version_from_dds}
\alias{strip_dotted_version_from_dds}
\title{Strip dotted version off of the rownames of a dds object}
\usage{
strip_dotted_version_from_dds(dds, force = FALSE)
}
\arguments{
\item{dds}{DESeqDataSet object}
}
\description{
Ensembl annotations frequently come with a dotted version number (e.g.,
ENSG00000001.3), but OrgDbs do not. This function removes the dotted
versions from the gene IDs of a DESeqDataSet object.
}
