% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{summarize.go}
\alias{summarize.go}
\title{Summarize and aggregate multiple GO results}
\usage{
summarize.go(ego, labels)
}
\arguments{
\item{ego}{List of clusterProfiler results objects}

\item{labels}{List of labels. For each name, a new column will be added and}
}
\value{
dataframe
}
\description{
Convert a list of GO analysis results (one per ontology) to a single
dataframe, with additional label column filled in with \code{label} and with
a fraction column.
}
