% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{summarize.kegg}
\alias{summarize.kegg}
\title{Summarize KEGG results}
\usage{
\method{summarize}{kegg}(ekg, labels)
}
\arguments{
\item{ekg}{Results from clusterProfiler KEGG enrichment}

\item{label}{Attach this label to the "label" column}
}
\value{
Dataframe
}
\description{
Summarize KEGG results and add \code{frac} and \code{label} columns
}
