% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize.res.list}
\alias{summarize.res.list}
\title{Combine everything in the results list into a single table}
\usage{
\method{summarize}{res.list}(res.list, alpha, lfc.thresh)
}
\arguments{
\item{res.list}{Named list of lists, where each sublist contains the following
names: c('res', 'dds', 'label'). "res" is a DESeqResults object,
"dds" is a DESeq object, and "label" is a nicer-looking
label to use.}
}
\value{
Dataframe
}
\description{
Combine everything in the results list into a single table
}
