% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_res_list}
\alias{summarize_res_list}
\title{Combine everything in the results list into a single table}
\usage{
summarize_res_list(res_list)
}
\arguments{
\item{res.list}{Named list of lists, where each sublist contains the
following names: c('res', 'dds', 'label'). "res" is a DESeqResults
object, "dds" is the indexing label into the dds_list object, and "label"
is a nicer-looking label to use. NOTE: backwards compatibility with older
versions of lcdb-wf depends on no dds.list object being passed.}
}
\value{
data.frame
}
\description{
Combine everything in the results list into a single table
}
