% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{top.plots}
\alias{top.plots}
\title{Plot normalized gene counts for the top N genes}
\usage{
top.plots(res, n, func, dds, add_cols = NULL, ...)
}
\arguments{
\item{n}{Number of top genes to plot}

\item{func}{Plotting function to call on each gene. Signature is func(gene,
dds, label=label, ...)}

\item{dds}{DESeq2 object}

\item{...}{Additional arguments that are passed to \code{func}}

\item{sorted}{DESeq2 results object}

\item{label}{Vector of column names in \code{res} from which to add a label to
the gene (e.g., c('symbol', 'alias'))}
}
\value{
Side effect is to create plot
}
\description{
Plot normalized gene counts for the top N genes
}
