% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{truncate}
\alias{truncate}
\title{Truncate the names of an enrichment results object}
\usage{
truncate(obj, truncate_to)
}
\arguments{
\item{obj}{DOSE::enrichResult object}

\item{truncate_to}{Max number of characters in a label}
}
\value{
enrichResult object with labels truncated
}
\description{
Truncate the names of an enrichment results object
}
