% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vargenes_heatmap}
\alias{vargenes_heatmap}
\title{Plot heatmap of most varying genes}
\usage{
vargenes_heatmap(rld, cols_for_grouping, n = 50)
}
\arguments{
\item{rld}{DESeqTransform object, typically output from running rlog()}

\item{n}{Number of genes to include}

\item{colData}{Dataframe of metadata, used for annotating heatmap}
}
\value{
Side effect is to plot the heatmap
}
\description{
Plot heatmap of most varying genes
}
