% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write.clusterprofiler.results}
\alias{write.clusterprofiler.results}
\title{Writes out original and split clusterprofiler results}
\usage{
write.clusterprofiler.results(res, cprof.folder, label)
}
\arguments{
\item{res}{clusterProfiler results}

\item{cprof.folder}{Directory in which to save files. Will be created if needed.}

\item{Label}{to use for the results. Will generate a filename
cprof.folder/label.txt and cprof.folder/label_split.txt}
}
\value{
List of the two files that are created on disk.
}
\description{
Writes out original and split clusterprofiler results
}
