% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplots.R
\name{write.upset.plot.results}
\alias{write.upset.plot.results}
\title{Save TSVs and PDFs of UpSet plot. Saves the current plot to PDF, so should
be called immediately after plotting.}
\usage{
write.upset.plot.results(lldf, label)
}
\arguments{
\item{lldf}{Dataframe of 1 and 0 from creating UpSet plots}

\item{label}{Label to use when constructing filenames}
}
\description{
Save TSVs and PDFs of UpSet plot. Saves the current plot to PDF, so should
be called immediately after plotting.
}
