% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{build_results_tabs}
\alias{build_results_tabs}
\title{Build tabbed output for results}
\usage{
build_results_tabs(res_list, dds_list, config, text)
}
\arguments{
\item{res_list}{List of results objects}

\item{dds_list}{List of dds objects}

\item{config}{Config object}

\item{text}{Text config object}
}
\value{
NULL, but creates the plots and tabs as a side effect.
}
\description{
This function outputs the RMarkdown and plots for building nested tabs for
each set of results. Make sure this is called within a \code{results="asis"}
chunk.
}
