% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{DESeqDataSetFromSalmon}
\alias{DESeqDataSetFromSalmon}
\title{Load Salmon quantification data into a DESeq object}
\usage{
DESeqDataSetFromSalmon(sampleTable, design, ignoreRank = FALSE, ...)
}
\arguments{
\item{sampleTable}{data.frame containing at least "salmon.path" column}

\item{design}{Model used for creating DESeq object}
}
\value{
DESeq object with transcript-level counts

Additional args are passed to DESeq2::DESeqDataSetFromMatrix.
}
\description{
Load Salmon quantification data into a DESeq object
}
