% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{enrich_list_lapply}
\alias{enrich_list_lapply}
\title{Apply a function to a nested enrichResults list}
\usage{
enrich_list_lapply(enrich_list, func, send_names = FALSE, ...)
}
\arguments{
\item{enrich_list}{Nested list. First level keys are results names. Second
level keys are directions (up, down, changed). Third level keys are
ontology labels ("ont").}

\item{func}{Function to apply to every identified enrichResult object}

\item{send_names}{If TRUE, \code{func} will also be provided with arguments for
name, direction, and ont. This can be useful if the function needs to
behave differently depending on the object.}
}
\value{
A list of the same shape (same nesting, same keys) but with the
"leaves" replaced with whatever \code{func} returns.
}
\description{
Apply a function to a nested enrichResults list
}
\details{
The results from many functional enrichment runs across many contrasts can
be tedious to work with. This function can be used to apply a function to
each "leaf" enrichResults object in the list.

The list is expected to have the structure:

list(
contrast_name=list(
direction=list(
ontology_name=enrichResults)
)
)

That is, \code{enrich_list[["ko.vs.wt"]][["up"]][["BP"]]} will get the GO
Biological Process enrichment results from the upregulated genes of contrast
"ko.vs.wt".
}
