% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{get_go_descriptions}
\alias{get_go_descriptions}
\title{Extract all GO IDs and their respective descriptions.}
\usage{
get_go_descriptions()
}
\value{
Two-column data.frame, GO accession in first column and description
in the second.
}
\description{
This is used, e.g., as a TERM2NAME dataframe to provide to
clusterProfiler::enricher or clusterProfiler::GSEA.
}
