% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make.dds.list}
\alias{make.dds.list}
\title{Make a list of dds objects}
\usage{
make.dds.list(
  deseq_obj_list,
  salmon.files = NULL,
  combine.by = FALSE,
  remove.version = TRUE,
  ...
)
}
\arguments{
\item{deseq_obj_list}{A named list of lists. Each list should have two to three elements:
1 should be a sampletable, 2 should be a design. The 3rd is an optional list of arguments
to pass to the DESeqDataSet constructor.}

\item{salmon.files}{txi object from tximport if transcript-level counts desired; NULL forces featurecounts;
shared by all dds.}

\item{combine.by}{The column to collapse technical replicates by.}

\item{remove.version}{Boolean to decide whether to remove version information}
}
\value{
A list of dds objects.
}
\description{
This function takes a list of sample, design pairs, plus other info, to output
a list of DESeq object.
}
