% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{make_results}
\alias{make_results}
\title{Convenience function for building contrasts}
\usage{
make_results(dds_name, label, dds_list = NULL, ...)
}
\arguments{
\item{dds_name}{String key into dds_list. Will be used to extract the actual
dds object to provide to results()}

\item{label}{Label to describe this contrast which will be used in headings.}

\item{dds_list}{List of dds objects. If NULL, then look in the global
environment for an object called "dds_list" and use that.}

\item{...}{Additional arguments are passed to results() and lfcShrink(). If
"parallel" is not explicitly specified here, then look in the global env for
a variable called "config" and find the parallel config setting from there.}
}
\value{
List as expected by lcdbwf, with the following names:
\itemize{
\item "res" (the results object)
\item "dds" (the string value passed in that was used to retrieve the actual
dds object used for the results)
\item "label" (exactly as passed in)
}
}
\description{
This function first runs DESeq2::results() and then runs DESeq2::lfcShrink()
on that results object. Arguments in ... are passed through to results() and
then to lfcShrink(), using formals() to detect which function accepts which
arguments. Those arguments accepted by both (contrast, lfcThreshold, format,
saveCols, parallel, BPPARAM) are passed to both.

This function is useful for ensuring that the dds object used when creating
the results is the same one that is labeled, so that when the dds_list and
res_list are composed together things match up nicely.

This function also adds the shrinkage type as an additional item to the
metadata of the results object so it can be reported.

NOTE: this function expects \code{dds_list} and \code{config} objects to be available
in the global environment.
}
