% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{make_results}
\alias{make_results}
\title{Convenience function for building contrasts using the data structures
expected by lcdbwf.}
\usage{
make_results(dds_name, label, ...)
}
\arguments{
\item{dds_name}{String key into dds_list}

\item{label}{Label or description to describe this contrast}

\item{...}{Additional arguments are passed to lfcShrink}

\item{contrast}{Any contrast supported by DESeq2::lfcShrink}
}
\value{
List with names "res", "dds", and "label".
}
\description{
NOTE: this function makes some assumptions about the global environment. For
example, it looks for the config object there, and expects a list called
"dds_list" to exist as well. Currently the config is only used for
parallelization.
}
\details{
This function is useful for ensuring that the dds object used when creating
the results is the same one that is labeled, so that when the dds_list and
res_list are composed together things match up nicely.
}
