% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize.res.list}
\alias{summarize.res.list}
\title{Combine everything in the results list into a single table}
\usage{
\method{summarize}{res.list}(res.list, alpha, lfc.thresh, dds.list = NULL)
}
\arguments{
\item{res.list}{Named list of lists, where each sublist contains the following
names: c('res', 'dds', 'label'). "res" is a DESeqResults object,
"dds" is either the indexing label for the dds.list object or
the DESeq object, and "label" is a nicer-looking
label to use. NOTE: backwards compatibility with older versions
of lcdb-wf depends on no dds.list object being passed.}
}
\value{
Dataframe
}
\description{
Combine everything in the results list into a single table
}
